+++ services/web/app/storage/api/misc.py
@@ -52,15 +52,29 @@ def storage_transfer_rack_to_shelf(tokenuser: UserAccount):
     except ValidationError as err:
         return validation_error_response(err)
 
-    ets = EntityToStorage.query.filter_by(
+    #ets = EntityToStorage.query.filter_by(
+    #    rack_id=values["rack_id"], storage_type="BTS"
+    #).first()
+    etss = EntityToStorage.query.filter_by(
         rack_id=values["rack_id"], storage_type="BTS"
-    ).first()
+    ).order_by(EntityToStorage.removed.asc()).all()
 
-    if ets is not None:
-        ets.box_id = None
-        ets.shelf_id = values["shelf_id"]
-        ets.editor_id = tokenuser.id
-        ets.storage_type = "BTS"
+
+    if len(etss)>0:
+        n=0
+        for ets in etss:
+            n = n+1
+            if n==1:
+                #ets.rack_id = values["rack_id"]
+                ets.shelf_id = values["shelf_id"]
+                ets.storage_type = "BTS"
+                ets.removed = False
+                ets.update({"editor_id": tokenuser.id})
+                db.session.add(ets)
+            else:
+
+                ets.update({"editor_id": tokenuser.id})
+                db.session.delete(ets)
 
     else:
         ets = EntityToStorage(**rack_to_shelf_result)
@@ -100,31 +114,37 @@ def storage_transfer_racks_to_shelf(tokenuser: UserAccount):
         except ValidationError as err:
             return validation_error_response(err)
 
-        # etss = EntityToStorage.query.filter_by(rack_id=values["rack_id"],
-        #             storage_type='BTS', removed=True).all()
-        # if len(etss)>0:
-        #     try:
-        #         for ets in etss:
-        #             # ets.removed = True
-        #             # ets.editor_id = tokenuser.id
-        #             # ets.updated_on = func.now()
-        #             # db.session.add(ets)
-        #             db.session.delete(ets)
-        #
-        #         db.session.flush()
-        #         print('okok')
-        #     except Exception as err:
-        #         return transaction_error_response(err)
+        etss = EntityToStorage.query.filter_by(rack_id=values["rack_id"],
+                    storage_type='BTS').order_by(EntityToStorage.removed).all()
+        if len(etss)>0:
 
-        ets = EntityToStorage(**rack_to_shelf_result)
-        ets.author_id = tokenuser.id
-        ets.storage_type = "BTS"
-        try:
+            try:
+                n = 0
+                for ets in etss:
+                    n=n+1
+                    if n>1:
+                        ets.update({"editor_id": tokenuser.id})
+                        db.session.delete(ets)
+                    else:
+                        ets.removed = False
+                        ets.shelf_id = rack_values["shelf_id"]
+                        ets.update({"editor_id": tokenuser.id})
+                        db.session.add(ets)
+
+                db.session.flush()
+            except Exception as err:
+                return transaction_error_response(err)
+        else:
+            ets = EntityToStorage(**rack_to_shelf_result)
+            ets.author_id = tokenuser.id
+            ets.storage_type = "BTS"
             db.session.add(ets)
-            db.session.flush()
 
-        except Exception as err:
-            return transaction_error_response(err)
+            try:
+                db.session.flush()
+
+            except Exception as err:
+                return transaction_error_response(err)
 
         ucs = UserCart.query.filter_by(rack_id=rack_id, storage_type="RUC").all()
 
@@ -156,6 +176,7 @@ def storage_transfer_racks_to_shelf(tokenuser: UserAccount):
 
             try:
                 db.session.add(rk)
+                db.session.flush()
             except Exception as err:
                 return transaction_error_response(err)
 
@@ -208,37 +229,44 @@ def storage_transfer_sample_to_shelf(tokenuser: UserAccount):
     except ValidationError as err:
         return validation_error_response(err)
 
-    etss = EntityToStorage.query.filter(sample_id=values["sample_id"]).all()
+    etss = EntityToStorage.query.filter(sample_id=values["sample_id"]).order_by(EntityToStorage.removed).all()
     if len(etss) > 0:
         # warning, confirmation
         try:
+            n=0
             for ets in etss:
-                db.session.delete(ets)
+                n=n+1
+                if n>1:
+                    ets.update({"editor_i": tokenuser.id})
+                    db.session.delete(ets)
+                    continue
+                else:
+                    ets.rack_id = None
+                    ets.shelf_id = values["shelf_id"]
+                    ets.storage_type = "STS"
+                    ets.removed = False
+                    ets.update({'editor_id': tokenuser.id})
+                    db.session.add(ets)
+
+            db.session.flush()
         except Exception as err:
             return transaction_error_response(err)
 
-    #
-    # etss = EntityToStorage.query.filter_by(sample_id=values["sample_id"], storage_type='STS').all()
-    # if ets != None:
-    #     ets.box_id = None
-    #     ets.shelf_id = values["shelf_id"]
-    #     ets.editor_id = tokenuser.id
-    #     ets.updated_on = func.now()
-    #     ets.storage_type = "STS"
-
-    ets = EntityToStorage(
-        sample_id=values["sample_id"],
-        shelf_id=values["shelf_id"],
-        storage_type="STS",
-        entry=values["entry"],
-        entry_datetime=values["entry_datetime"],
-        author_id=tokenuser.id,
-    )
-    try:
+    else:
+        ets = EntityToStorage(
+            sample_id=values["sample_id"],
+            shelf_id=values["shelf_id"],
+            storage_type="STS",
+            entry=values["entry"],
+            entry_datetime=values["entry_datetime"],
+            author_id=tokenuser.id,
+        )
         db.session.add(ets)
-        db.session.flush()
-    except Exception as err:
-        return transaction_error_response(err)
+
+        try:
+            db.session.flush()
+        except Exception as err:
+            return transaction_error_response(err)
 
     usercart = UserCart.query.filter_by(
         sample_id=values["sample_id"], author_id=tokenuser.id
@@ -278,7 +306,7 @@ def storage_transfer_samples_to_shelf(tokenuser: UserAccount):
 
         etss = (EntityToStorage.query
                 .filter_by(sample_id=values["sample_id"])
-                .order_by(EntityToStorage.removed.asc())
+                .order_by(EntityToStorage.removed)
                 .all()
                 )
 
@@ -301,6 +329,7 @@ def storage_transfer_samples_to_shelf(tokenuser: UserAccount):
                         ets.update({"editor_id":tokenuser.id})
                         db.session.add(ets)
                     else:
+                        ets.update({"editor_id": tokenuser.id})
                         db.session.delete(ets)
                         continue
